gulp supports es6 now

+++ gulpfile.babel.js
@@ -11,19 +11,17 @@ const webpack = require('webpack-stream');
 const spawn = require('child_process').spawn;
 const seq = require('run-sequence');
 const uglify = require('gulp-uglify');
+const fs = require('fs');
 
 let radarInstance;
 
 // Copy some theme stuff over
 gulp.task('assets', function () {
-  gulp.src('www/favicon.ico')
-    .pipe(gulp.dest('public/'));
+  gulp.src('www/favicon.ico').pipe(gulp.dest('public/'));
 });
 
 gulp.task('radar:restart', function () {
-  if (radarInstance) {
-    radarInstance.kill();
-  }
+  if (radarInstance) { radarInstance.kill(); }
   radarInstance = spawn('./bin/doplr.js', ['radar'], {
     stdio: 'inherit',
     env: {},
@@ -41,7 +39,7 @@ process.on('exit', function() {
   if (radarInstance) { radarInstance.kill(); }
 });
 
-gulp.task('weathergirl:start', function () {
+gulp.task('reloadproxy:start', function () {
   connect.server({
     root: 'public',
     port: 8080,
@@ -81,7 +79,7 @@ gulp.task('webpack', function () {
 });
 
 gulp.task('less', function () {
-  return gulp.src('www/index.less')
+  gulp.src('www/index.less')
     .pipe(less({
       compress: process.env.COMPRESS || false,
       rootpath: '/'
@@ -96,18 +94,18 @@ gulp.task('less', function () {
 });
 
 gulp.task('semantic:setup', function () {
-  // Setup semantic node_module for use
-  return gulp.src('node_modules/semantic-ui/src/_site/*/*')
-    .pipe(gulp.dest('node_modules/semantic-ui/src/site'));
-});
-
-gulp.task('semantic:ui', function () {
   // Theme config
   gulp.src('www/theme.config')
     .pipe(gulp.dest('node_modules/semantic-ui/src'));
   // Fonts
   gulp.src('node_modules/semantic-ui/src/themes/default/assets/fonts/*')
     .pipe(gulp.dest('public/themes/default/assets/fonts/'));
+  // Setup semantic node_module for use
+  return gulp.src('node_modules/semantic-ui/src/_site/*/*')
+    .pipe(gulp.dest('node_modules/semantic-ui/src/site'));
+});
+
+gulp.task('semantic:ui', function () {
   // Compile semantic into a vendor.css bundle
   gulp.src('node_modules/semantic-ui/src/semantic.less')
     .pipe(less({
@@ -124,7 +122,7 @@ gulp.task('semantic:ui', function () {
 });
 
 gulp.task('jade', function () {
-  return gulp.src('www/index.jade')
+  gulp.src('www/index.jade')
     .pipe(jade({
       locals: {
         assetURL: '/'
@@ -138,36 +136,42 @@ gulp.task('jade', function () {
     .pipe(connect.reload());
 });
 
-const defaultTasks = ['less', 'jade', 'webpack', 'assets'];
+const semanticIsSetup = fs.existsSync('node_modules/semantic-ui/src/site');
+const defaultTasks = ['semantic:ui', 'less', 'jade', 'webpack', 'assets'];
 gulp.task('default', function () {
-  seq('semantic:setup', 'semantic:ui', defaultTasks);
+  if (semanticIsSetup) {
+    seq(defaultTasks);
+  } else {
+    seq('semantic:setup', defaultTasks);
+  }
 });
 
 gulp.task('watch', function () {
-  seq(
-    'semantic:setup',
-    'semantic:ui',
-    defaultTasks,
-    ['radar:restart', 'weathergirl:start'],
-    function () {
-      // LESS FILES
-      watch(['www/index.less'], function () {
-        seq(['less']);
+  const tasks = [];
+  if (!semanticIsSetup) {
+    tasks.push('semantic:setup');
+  }
+  tasks.push(defaultTasks);
+  tasks.push(['radar:restart', 'reloadproxy:start']);
+  tasks.push(function () {
+    // LESS FILES
+    watch(['www/index.less'], function () {
+      seq(['less']);
+    });
+    // JADE FILES
+    watch(['www/index.jade'], function () {
+      seq(['jade']);
+    });
+    // FRONTEND: JS AND JSX
+    watch(['www/*.jsx', 'www/*.js', 'www/partials/*.jsx', 'www/partials/*/*.jsx'], function () {
+      seq(['webpack']);
+    });
+    // DOPLR LIB: JS
+    watch(['lib/*.js', 'lib/sweep/*.js'], function () {
+      seq(['radar:restart'], function () {
+        connect.reload();
       });
-      // JADE FILES
-      watch(['www/index.jade'], function () {
-        seq(['jade']);
-      });
-      // FRONTEND: JS AND JSX
-      watch(['www/*.jsx', 'www/*.js', 'www/partials/*.jsx', 'www/partials/*/*.jsx'], function () {
-        seq(['webpack']);
-      });
-      // DOPLR LIB: JS
-      watch(['lib/*.js', 'lib/sweep/*.js'], function () {
-        seq(['radar:restart'], function () {
-          connect.reload();
-        });
-      });
-    }
-  );
+    });
+  });
+  seq(...tasks);
 });
